# api/openapi.yaml
openapi: "3.1.0"
info:
  title: OurFamRoots API
  version: "1.0.0"
  description: |
    ## OurFamRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
...
  schemas:
...
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://ourfamroots.com/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
...
    NamespaceInvitation:
      type: object
      description: >
        Invitation transferring an existing Global-namespace user account into a
        specific namespace with a chosen role. Accepting updates the user's
        tenant_id/app_role in place — it never creates a second account.
      required: [id, tenant_id, namespace_name, inviter_name, invitee_email, role, status, expires_at, created_at]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        namespace_name: { type: string }
        inviter_name: { type: string }
        invitee_email: { type: string, format: email }
        role:
          type: string
          description: One of `ADMIN`, `STANDARD`, `AUDITOR` — the app role the invitee will have once accepted.
        status:
          type: string
          description: One of `PENDING`, `ACCEPTED`, `DECLINED`, `EXPIRED`, `REVOKED`.
        expires_at:
          type: string
          format: date-time
          description: 168 hours (7 days) after creation.
        created_at: { type: string, format: date-time }
...
  responses:
...
    401Unauthorized:
      description: Unauthorized — missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    403Forbidden:
      description: Forbidden — valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    422UnprocessableEntity:
      description: Unprocessable Entity — validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /admin/namespaces/{namespaceId}/invitations:
    parameters:
      - name: namespaceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Admin, Namespaces]
      operationId: createNamespaceInvitation
      summary: Invite a Global-namespace user into this namespace
      description: |
        Requires the `SUPER_ADMIN`, or an `ADMIN` of the target namespace.

        The invitee must already be an existing user of the Global namespace
        (found by email) — namespace invitations only ever move an existing
        account, they never create a new one. Cannot target the Global
        namespace itself.

        **Side effects:**
        - Invitation record created with a 168-hour (7-day) expiry.
        - Invitation email sent to the invitee (best-effort — failure does
          not roll back the invitation).
        - In-app notification + Web Push sent to the invitee (best-effort).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invitee_email, role]
              properties:
                invitee_email: { type: string, format: email }
                role:
                  type: string
                  enum: [ADMIN, STANDARD, AUDITOR]
                message: { type: string, maxLength: 500, nullable: true }
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamespaceInvitation"
        "400":
          description: No Global namespace is configured, or the target namespace is the Global namespace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          description: Namespace not found, or no Global-namespace user found with that email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
  /admin/namespaces/{namespaceId}/invitations/{invitationId}:
    parameters:
      - name: namespaceId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: invitationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Admin, Namespaces]
      operationId: revokeNamespaceInvitation
      summary: Revoke a pending namespace invitation
      description: |
        Requires the `SUPER_ADMIN`, or an `ADMIN` of the target namespace.
        Only a `PENDING` invitation can be revoked.
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Invitation revoked — no body
        "400":
          description: Only a pending invitation can be revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
  /namespace-invitations/pending:
    get:
      tags: [Namespaces]
      operationId: listPendingNamespaceInvitations
      summary: List pending namespace invitations addressed to the current user
      description: Requires a verified user account.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Pending invitations for this account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamespaceInvitation"
        "401":
          $ref: "#/components/responses/401Unauthorized"
  /namespace-invitations/{token}/accept:
    parameters:
      - name: token
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Namespaces]
      operationId: acceptNamespaceInvitation
      summary: Accept a namespace invitation — transfers your account into that namespace
      description: |
        Requires a verified user account, and the invitation must be
        addressed to that same account.

        **Side effects:**
        - Updates the caller's `tenant_id`/`app_role` in place (does not
          create a second account).
        - Drops stale tree memberships from the old namespace and grants the
          new namespace's globally-shared trees.
        - Marks the invitation `ACCEPTED`.
        - Revokes all of the caller's refresh tokens/sessions — the JWT's
          `tid`/`role` claims are now stale, so this forces a re-login
          everywhere.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant_id: { type: string, format: uuid }
                  app_role: { type: string, enum: [ADMIN, STANDARD, AUDITOR] }
        "400":
          description: Invitation is no longer pending, or has expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          description: This invitation was addressed to a different account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/404NotFound"

# ============================================================
# USER GROUPS (Admin)
# ============================================================
...